+++ spec/spec_helper.rb
@@ -35,6 +35,8 @@ RSpec.configure do |config|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
+  config.filter_run_when_matching :focus
+
   config.mock_with :rspec do |mocks|
     mocks.verify_partial_doubles = true
   end